perf(docs): pre-render markdown to HTML at build time
Docs were processed through unified+shiki on every serverless cold start
(~2-5s for 18 files). Now pre-rendered at build time into a static JSON
file that's bundled into the app. Runtime is a simple array lookup.

- scripts/build-docs.ts: build-time processor (remark → rehype → shiki)
- query/docs.ts: imports pre-rendered JSON, zero runtime processing
- Generated docs.json is gitignored (built in CI)
